% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicedownApp-package.R
\docType{package}
\name{juicedownApp-package}
\alias{juicedownApp}
\alias{juicedownApp-package}
\title{juicedownApp: Small Shiny App for CMS Pages}
\description{
Small Shiny App That Helps You Convert a Markdown Document into an HTML Fragment that's Ready for Copy-and-Pasting into Moodle.
}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID}) [copyright holder]

}
\keyword{internal}
