% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up.R
\name{up}
\alias{up}
\title{Launch shiny app}
\usage{
up(allow_r = TRUE, ...)
}
\arguments{
\item{allow_r}{logical. If set to FALSE, code chunks are converted into a
verbatim code block during the conversion process.}

\item{...}{Not used yet.}
}
\value{
Launch a shiny app.
}
\description{
Launch shiny app
}
